@@ -1,10 +1,17 @@
-const List<Map<String, String>> usuarios = [
+import 'dart:math'; // Importa dart:math para generar números aleatorios
+import 'package:unidad_cumplimiento/src/datos_proceso.dart'; // Importa datos_proceso.dart para usar la lista "organismos"
+
+final Random random = Random();
+
+final List<Map<String, String>> usuarios = [
   {
     'nombre': 'Juan Pérez',
     'userImagen': 'https://randomuser.me/api/portraits/men/1.jpg',
     'estado': 'Activo',
     'profesion': 'Ingeniero',
     'correoElectronico': 'juan.perez@example.com',
+    'celular': '3001234567',
+    'organismo': organismos[random.nextInt(organismos.length)], // Asigna un organismo aleatorio
   },
   {
     'nombre': 'María García',
@@ -12,6 +19,8 @@ const List<Map<String, String>> usuarios = [
     'estado': 'Inactivo',
     'profesion': 'Doctora',
     'correoElectronico': 'maria.garcia@example.com',
+    'celular': '3012345678',
+    'organismo': organismos[random.nextInt(organismos.length)], // Asigna un organismo aleatorio
   },
   {
     'nombre': 'Carlos López',
@@ -19,6 +28,8 @@ const List<Map<String, String>> usuarios = [
     'estado': 'Activo',
     'profesion': 'Abogado',
     'correoElectronico': 'carlos.lopez@example.com',
+    'celular': '3023456789',
+    'organismo': organismos[random.nextInt(organismos.length)], // Asigna un organismo aleatorio
   },
   {
     'nombre': 'José González',
@@ -26,6 +37,8 @@ const List<Map<String, String>> usuarios = [
     'estado': 'Activo',
     'profesion': 'Contador',
     'correoElectronico': 'jose.gonzalez@example.com',
+    'celular': '3034567890',
+    'organismo': organismos[random.nextInt(organismos.length)], // Asigna un organismo aleatorio
   },
   {
     'nombre': 'Marta Sánchez',
@@ -33,6 +46,8 @@ const List<Map<String, String>> usuarios = [
     'estado': 'Inactivo',
     'profesion': 'Psicóloga',
     'correoElectronico': 'marta.sanchez@example.com',
+    'celular': '3045678901',
+    'organismo': organismos[random.nextInt(organismos.length)], // Asigna un organismo aleatorio
   },
   {
     'nombre': 'David Ramírez',
@@ -40,6 +55,8 @@ const List<Map<String, String>> usuarios = [
     'estado': 'Activo',
     'profesion': 'Desarrollador',
     'correoElectronico': 'david.ramirez@example.com',
+    'celular': '3056789012',
+    'organismo': organismos[random.nextInt(organismos.length)], // Asigna un organismo aleatorio
   },
   {
     'nombre': 'Sara Torres',
@@ -47,6 +64,8 @@ const List<Map<String, String>> usuarios = [
     'estado': 'Inactivo',
     'profesion': 'Diseñadora',
     'correoElectronico': 'sara.torres@example.com',
+    'celular': '3067890123',
+    'organismo': organismos[random.nextInt(organismos.length)], // Asigna un organismo aleatorio
   },
   {
     'nombre': 'Jorge Díaz',
@@ -54,6 +73,8 @@ const List<Map<String, String>> usuarios = [
     'estado': 'Activo',
     'profesion': 'Analista',
     'correoElectronico': 'jorge.diaz@example.com',
+    'celular': '3078901234',
+    'organismo': organismos[random.nextInt(organismos.length)], // Asigna un organismo aleatorio
   },
   {
     'nombre': 'Elena Ruiz',
@@ -61,6 +82,8 @@ const List<Map<String, String>> usuarios = [
     'estado': 'Inactivo',
     'profesion': 'Investigadora',
     'correoElectronico': 'elena.ruiz@example.com',
+    'celular': '3089012345',
+    'organismo': organismos[random.nextInt(organismos.length)], // Asigna un organismo aleatorio
   },
   {
     'nombre': 'Pedro Morales',
@@ -68,6 +91,8 @@ const List<Map<String, String>> usuarios = [
     'estado': 'Activo',
     'profesion': 'Gerente',
     'correoElectronico': 'pedro.morales@example.com',
+    'celular': '3090123456',
+    'organismo': organismos[random.nextInt(organismos.length)], // Asigna un organismo aleatorio
   },
   {
     'nombre': 'Isabel Ortiz',
@@ -75,6 +100,8 @@ const List<Map<String, String>> usuarios = [
     'estado': 'Inactivo',
     'profesion': 'Consultora',
     'correoElectronico': 'isabel.ortiz@example.com',
+    'celular': '3101234567',
+    'organismo': organismos[random.nextInt(organismos.length)], // Asigna un organismo aleatorio
   },
   {
     'nombre': 'Ricardo Castro',
@@ -82,5 +109,7 @@ const List<Map<String, String>> usuarios = [
     'estado': 'Activo',
     'profesion': 'Administrador',
     'correoElectronico': 'ricardo.castro@example.com',
+    'celular': '3112345678',
+    'organismo': organismos[random.nextInt(organismos.length)], // Asigna un organismo aleatorio
   },
 ];
