@@ -4,9 +4,9 @@ class SecondaryAppBar extends StatelessWidget implements PreferredSizeWidget {
   final String title;
 
   const SecondaryAppBar({
-    Key? key,
+    super.key,
     required this.title,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
